#1013 Extract stories helper (#1041)
* Extract get_stories to stories helper

* Refactor StoriesHelper

* Avoid including StoriesHelper in ApplicationController
