+++ app/controllers/application_controller.rb
@@ -86,7 +86,7 @@ class ApplicationController < ActionController::Base
   helper_method :avatar_url, :viewer_of?,
                 :are_allies?, :get_uid, :most_focus,
                 :tag_usage, :can_notify, :if_not_signed_in,
-                :generate_comment, :get_stories, :moments_stats
+                :generate_comment, :moments_stats
 
   def if_not_signed_in
     return if user_signed_in?
@@ -164,39 +164,6 @@ class ApplicationController < ActionController::Base
       end
     end
     result
-  end
-  # rubocop:enable MethodLength
-
-  # rubocop:disable MethodLength
-  def get_stories(user, include_allies)
-    if user.id == current_user.id
-      my_moments = user.moments.all.recent
-      my_strategies = user.strategies.all.recent
-    end
-
-    if include_allies && user.id == current_user.id
-      allies = user.allies_by_status(:accepted)
-      allies.each do |ally|
-        my_moments += user_stories(ally, 'moments')
-        my_strategies += user_stories(ally, 'strategies')
-      end
-    elsif !include_allies && user.id != current_user.id
-      my_moments = user_stories(user, 'moments')
-      my_strategies = user_stories(user, 'strategies')
-    end
-
-    moments = Moment.where(id: my_moments.map(&:id)).order(created_at: :desc)
-    strategies = Strategy.where(id: my_strategies.map(&:id))
-                         .order(created_at: :desc)
-
-    stories =
-      if moments.count.positive?
-        moments.zip(strategies).flatten.compact
-      else
-        strategies.compact
-      end
-
-    stories.sort_by(&:created_at).reverse!
   end
   # rubocop:enable MethodLength
 
@@ -307,19 +274,6 @@ class ApplicationController < ActionController::Base
 
   def user_moments(user_id)
     Moment.where(user_id: user_id)
-  end
-
-  def user_stories(user, collection)
-    case collection
-    when 'moments'
-      query = Moment.published
-    when 'strategies'
-      query = Strategy.published
-    end
-
-    query.where(user_id: user.id).all.recent.map do |story|
-      story if story.viewers.include?(current_user.id)
-    end.compact
   end
 end
 # rubocop:enable ClassLength
